Add a test case for not throwing if a key is missing
The new test case verifies the current behavior that no exceptions are
thrown if a key mentioned in a template is not found.

The Catch2 TEST_CASE macro was used directly instead of the MSTCH_TEST
or other macros because this test case does not verify output of the
mstch::render with a given input, what MSTCH_TEST and friends are for.

Signed-off-by: Hannu Lounento <hannu.lounento@vaisala.com>